// generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Role = {
  Customer: 'Customer',
  Admin: 'Admin',
  Staff: 'Staff'
} as const

export type Role = (typeof Role)[keyof typeof Role]


export const OrderStatus = {
  Pending: 'Pending',
  Preparing: 'Preparing',
  Ready: 'Ready',
  Completed: 'Completed',
  Cancelled: 'Cancelled'
} as const

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus]


export const PaymentStatus = {
  Unpaid: 'Unpaid',
  Paid: 'Paid'
} as const

export type PaymentStatus = (typeof PaymentStatus)[keyof typeof PaymentStatus]